+++ sqlite3_ext_macro/src/lib.rs
@@ -369,7 +369,7 @@ pub fn sqlite3_ext_doctest_impl(item: TokenStream) -> TokenStream {
         struct Cursor {}
         impl ::sqlite3_ext::vtab::VTabCursor for Cursor {
             type ColumnType = ();
-            fn filter(&mut self, _: usize, _: Option<&str>, _: &mut [&mut ValueRef]) -> std::result::Result<(), ::sqlite3_ext::Error> { todo!() }
+            fn filter(&mut self, _: i32, _: Option<&str>, _: &mut [&mut ValueRef]) -> std::result::Result<(), ::sqlite3_ext::Error> { todo!() }
             fn next(&mut self) -> std::result::Result<(), ::sqlite3_ext::Error> { todo!() }
             fn eof(&self) -> bool { todo!() }
             fn column(&self, _: usize) { todo!() }
